% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as_arrow_table}
\alias{as_arrow_table}
\title{Convert an lb_result to an Arrow Table}
\usage{
as_arrow_table(x, ...)
}
\arguments{
\item{x}{An \code{lb_result} object.}

\item{...}{Unused.}
}
\value{
An \link[arrow:Table-class]{arrow::Table} object.
}
\description{
Convert an lb_result to an Arrow Table
}
