% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as_tbl_graph}
\alias{as_tbl_graph}
\title{Convert an lb_result to a tbl_graph (tidygraph)}
\usage{
as_tbl_graph(x, ...)
}
\arguments{
\item{x}{An \code{lb_result} object.}

\item{...}{Unused.}
}
\value{
A \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} object.
}
\description{
Convert an lb_result to a tbl_graph (tidygraph)
}
