% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as_tibble.lb_result}
\alias{as_tibble.lb_result}
\title{Convert an lb_result to a tibble}
\usage{
as_tibble.lb_result(x, ...)
}
\arguments{
\item{x}{An \code{lb_result} object.}

\item{...}{Passed to \code{\link[=as.data.frame.lb_result]{as.data.frame.lb_result()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Convert an lb_result to a tibble
}
