% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as_tibble}
\alias{as_tibble}
\title{Convert a LadybugDB result to a tibble}
\usage{
as_tibble(x, ...)
}
\arguments{
\item{x}{An \code{lb_result}.}

\item{...}{Passed to \code{\link[=as.data.frame]{as.data.frame()}}.}
}
\value{
A tibble.
}
\description{
Convert a LadybugDB result to a tibble
}
