% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.lb_check_conn}
\alias{.lb_check_conn}
\title{Assert that an object is an lb_connection}
\usage{
.lb_check_conn(conn)
}
\arguments{
\item{conn}{Object to check.}
}
\description{
Assert that an object is an lb_connection
}
\keyword{internal}
