% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{format.lb_interval}
\alias{format.lb_interval}
\title{Format a LadybugDB interval}
\usage{
\method{format}{lb_interval}(x, ...)
}
\arguments{
\item{x}{An \code{lb_interval} value.}

\item{...}{Unused.}
}
\description{
Format a LadybugDB interval
}
