% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\docType{class}
\name{LadybugConnection-class}
\alias{LadybugConnection-class}
\title{LadybugDB DBI connection}
\description{
LadybugDB DBI connection
}
\section{Slots}{

\describe{
\item{\code{database}}{Native database object.}

\item{\code{connection}}{Native connection object.}

\item{\code{state}}{Mutable connection state.}

\item{\code{bigint}}{INT64 conversion policy.}
}}

