% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{ladybugdb_is_installed}
\alias{ladybugdb_is_installed}
\title{Check whether LadybugDB is available}
\usage{
ladybugdb_is_installed()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Returns \code{TRUE} if the native LadybugDB C library is loaded and functional.
}
\examples{
ladybugdb_is_installed()

}
