% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{ladybugdb_version}
\alias{ladybugdb_version}
\title{Return the LadybugDB C library version string}
\usage{
ladybugdb_version()
}
\value{
A character string, e.g. \code{"0.20.4"}.
}
\description{
Return the LadybugDB C library version string
}
\examples{
ladybugdb_version()

}
