% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\docType{class}
\name{LadybugResult-class}
\alias{LadybugResult-class}
\title{LadybugDB DBI result}
\description{
LadybugDB DBI result
}
\section{Slots}{

\describe{
\item{\code{connection}}{Owning DBI connection.}

\item{\code{state}}{Mutable result state.}

\item{\code{statement}}{Original Cypher statement.}
}}

