% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{lb_begin}
\alias{lb_begin}
\alias{lb_commit}
\alias{lb_rollback}
\title{Transaction control}
\usage{
lb_begin(conn)

lb_commit(conn)

lb_rollback(conn)
}
\arguments{
\item{conn}{An \code{lb_connection}.}
}
\value{
Invisible \code{conn}.
}
\description{
Transaction control
}
