% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{lb_connection}
\alias{lb_connection}
\title{Open a connection to a LadybugDB database}
\usage{
lb_connection(database, num_threads = NULL)
}
\arguments{
\item{database}{An \code{lb_database} object created by \code{\link[=lb_database]{lb_database()}}.}

\item{num_threads}{Integer. Number of threads LadybugDB may use. \code{NULL}
leaves the default (typically all available cores).}
}
\value{
An object of class \code{lb_connection}.
}
\description{
Open a connection to a LadybugDB database
}
\examples{
\dontrun{
db   <- lb_database(":memory:")
conn <- lb_connection(db)
}

}
