% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{lb_fetch_arrow}
\alias{lb_fetch_arrow}
\title{Fetch a chunk as an Arrow Table}
\usage{
lb_fetch_arrow(result, n = Inf)
}
\arguments{
\item{result}{An \code{lb_result}.}

\item{n}{Maximum rows to fetch. \code{Inf} fetches all remaining rows.}
}
\value{
An \code{arrow::Table}.
}
\description{
Fetch a chunk as an Arrow Table
}
