% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{lb_next_result}
\alias{lb_next_result}
\title{Retrieve the next result from a multi-statement query}
\usage{
lb_next_result(result)
}
\arguments{
\item{result}{An \code{lb_result}.}
}
\value{
The next \code{lb_result}, or \code{NULL}.
}
\description{
Retrieve the next result from a multi-statement query
}
