% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{lb_query_summary}
\alias{lb_query_summary}
\title{Return query timing information}
\usage{
lb_query_summary(result)
}
\arguments{
\item{result}{An \code{lb_result}.}
}
\value{
A list with compilation and execution times in milliseconds.
}
\description{
Return query timing information
}
