% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{lb_result_info}
\alias{lb_result_info}
\title{Inspect result metadata}
\usage{
lb_result_info(result)
}
\arguments{
\item{result}{An \code{lb_result}.}
}
\value{
A named list.
}
\description{
Inspect result metadata
}
