% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{lb_set_timeout}
\alias{lb_set_timeout}
\alias{lb_interrupt}
\title{Configure and interrupt queries}
\usage{
lb_set_timeout(conn, milliseconds)

lb_interrupt(conn)
}
\arguments{
\item{conn}{An \code{lb_connection}.}

\item{milliseconds}{Non-negative timeout in milliseconds. Zero disables the
timeout.}
}
\value{
Invisible \code{conn}.
}
\description{
Configure and interrupt queries
}
