% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rladybugdb-package.R
\docType{package}
\name{rladybugdb-package}
\alias{rladybugdb}
\alias{rladybugdb-package}
\title{rladybugdb: Native R access to LadybugDB}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an R interface to LadybugDB, an embedded columnar graph database using the Cypher query language. LadybugDB was formerly known as Kuzu. This package uses native Rcpp bindings to the LadybugDB C library, requiring no Python runtime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadimaster65555/rladybugdb}
  \item Report bugs at \url{https://github.com/hadimaster65555/rladybugdb/issues}
}

}
\author{
\strong{Maintainer}: Raden Muhammad Hadi \email{hadimaster65555@gmail.com}

Authors:
\itemize{
  \item Claude (Anthropic AI assistant)
}

}
\keyword{internal}
