% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{ladybugdb_is_installed}
\alias{ladybugdb_is_installed}
\title{Check whether real_ladybug is importable}
\usage{
ladybugdb_is_installed()
}
\value{
\code{TRUE} if the Python package \code{real_ladybug} can be imported,
\code{FALSE} otherwise.
}
\description{
Check whether real_ladybug is importable
}
\examples{
ladybugdb_is_installed()

}
