% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{lb_connection}
\alias{lb_connection}
\title{Open a connection to a LadybugDB database}
\usage{
lb_connection(database, num_threads = NULL)
}
\arguments{
\item{database}{An \code{lb_database} object.}

\item{num_threads}{Positive integer execution thread limit, or \code{NULL}.}
}
\value{
An \code{lb_connection} object.
}
\description{
Open a connection to a LadybugDB database
}
